requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.0.5
psycopg2-binary
python-dotenv
gunicorn
